Fixing subtract vectors memory issues

+++ src/tests/subtract-two-vectors/SubtractTwoVectors.c
@@ -1,4 +1,4 @@
-void subtractTwoVectors(int *a, int *b, int *c, int n) {
+void subtractTwoVectors(double *a, double *b, double *c, int n) {
   for (int i = 0; i < n; i++) {
     c[i] = a[i] - b[i];
   }
